+++ routing/http/types/ndjson/records.go
@@ -2,6 +2,7 @@ package ndjson
 
 import (
 	"encoding/json"
+	"fmt"
 	"io"
 
 	"github.com/ipfs/boxo/routing/http/types"
@@ -9,7 +10,7 @@ import (
 )
 
 // NewRecordsIter returns an iterator that reads [types.Record] from the given [io.Reader].
-func NewRecordsIter(r io.Reader) iter.Iter[iter.Result[types.Record]] {
+func NewRecordsIter(r io.Reader) iter.ResultIter[types.Record] {
 	jsonIter := iter.FromReaderJSON[types.UnknownRecord](r)
 	mapFn := func(upr iter.Result[types.UnknownRecord]) iter.Result[types.Record] {
 		var result iter.Result[types.Record]
@@ -26,10 +27,8 @@ func NewRecordsIter(r io.Reader) iter.Iter[iter.Result[types.Record]] {
 				return result
 			}
 			result.Val = &prov
-		//lint:ignore SA1019 // ignore staticcheck
-		case types.SchemaBitswap:
-			//lint:ignore SA1019 // ignore staticcheck
-			var prov types.BitswapRecord
+		case types.SchemaAnnouncement:
+			var prov types.AnnouncementRecord
 			err := json.Unmarshal(upr.Val.Bytes, &prov)
 			if err != nil {
 				result.Err = err
@@ -42,32 +41,54 @@ func NewRecordsIter(r io.Reader) iter.Iter[iter.Result[types.Record]] {
 		return result
 	}
 
-	return iter.Map[iter.Result[types.UnknownRecord]](jsonIter, mapFn)
+	return iter.Map(jsonIter, mapFn)
 }
 
-// NewPeerRecordsIter returns an iterator that reads [types.PeerRecord] from the given [io.Reader].
-// Records with a different schema are safely ignored. If you want to read all records, use
+// NewAnnouncementRecordsIter returns an iterator that reads [types.AnnouncementRecord]
+// from the given [io.Reader]. Incompatible records result in an error within the iterator.
+// To read all records, use [NewRecordsIter] instead.
+func NewAnnouncementRecordsIter(r io.Reader) iter.ResultIter[*types.AnnouncementRecord] {
+	return newTypedRecords[*types.AnnouncementRecord](r, types.SchemaAnnouncement)
+}
+
+// NewAnnouncementResponseRecordsIter returns an iterator that reads
+// [types.AnnouncementResponseRecord] from the given [io.Reader]. Incompatible
+// records result in an error within the iterator. To read all records, use
 // [NewRecordsIter] instead.
-func NewPeerRecordsIter(r io.Reader) iter.Iter[iter.Result[*types.PeerRecord]] {
-	jsonIter := iter.FromReaderJSON[types.UnknownRecord](r)
-	mapFn := func(upr iter.Result[types.UnknownRecord]) iter.Result[*types.PeerRecord] {
-		var result iter.Result[*types.PeerRecord]
-		if upr.Err != nil {
-			result.Err = upr.Err
-			return result
-		}
-		switch upr.Val.Schema {
-		case types.SchemaPeer:
-			var prov types.PeerRecord
-			err := json.Unmarshal(upr.Val.Bytes, &prov)
-			if err != nil {
-				result.Err = err
+func NewAnnouncementResponseRecordsIter(r io.Reader) iter.ResultIter[*types.AnnouncementResponseRecord] {
+	return newTypedRecords[*types.AnnouncementResponseRecord](r, types.SchemaAnnouncementResponse)
+}
+
+// NewPeerRecordsIter returns an iterator that reads [types.PeerRecord] from the given
+// [io.Reader]. Incompatible records result in an error within the iterator. To read all records,
+// use [NewRecordsIter] instead.
+func NewPeerRecordsIter(r io.Reader) iter.ResultIter[*types.PeerRecord] {
+	return newTypedRecords[*types.PeerRecord](r, types.SchemaPeer)
+}
+
+func newTypedRecords[T any](r io.Reader, schema string) iter.ResultIter[T] {
+	return iter.Map(
+		NewRecordsIter(r),
+		func(upr iter.Result[types.Record]) iter.Result[T] {
+			var result iter.Result[T]
+			if upr.Err != nil {
+				result.Err = upr.Err
 				return result
 			}
-			result.Val = &prov
-		}
-		return result
-	}
 
-	return iter.Map[iter.Result[types.UnknownRecord]](jsonIter, mapFn)
+			if upr.Val.GetSchema() != schema {
+				result.Err = fmt.Errorf("unexpected schema %s, expected %s", upr.Val.GetSchema(), schema)
+				return result
+			}
+
+			val, ok := upr.Val.(T)
+			if !ok {
+				result.Err = fmt.Errorf("type incompatible with schema %s", schema)
+				return result
+			}
+
+			result.Val = val
+			return result
+		},
+	)
 }